pyproject.toml
[project]
name = "msig"
version = "0.1.0"
description = "Многопользовательская интерактивная игра для стримеров"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "email-validator>=2.2.0",
    "psycopg[binary]>=3.2.4",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
]

[tool.pytest.ini_options]
markers = [
    "interface_rules",
]
